Put reshape utils in one file
