+++ mlx/backend/gpu/CMakeLists.txt
@@ -2,4 +2,5 @@ target_sources(
   mlx
   PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/copy.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/primitives.cpp
+          ${CMAKE_CURRENT_SOURCE_DIR}/reshape.cpp
           ${CMAKE_CURRENT_SOURCE_DIR}/slicing.cpp)
